Removed services sub-package from models
