+++ src/main/java/com/exceptionless/exceptionlessclient/models/Event.java
@@ -3,9 +3,7 @@ package com.exceptionless.exceptionlessclient.models;
 import com.exceptionless.exceptionlessclient.models.base.Model;
 import com.exceptionless.exceptionlessclient.models.enums.EventPropertyKey;
 import com.exceptionless.exceptionlessclient.models.enums.EventTag;
-import com.exceptionless.exceptionlessclient.models.services.EnvironmentInfo;
-import com.exceptionless.exceptionlessclient.models.services.RequestInfo;
-import com.exceptionless.exceptionlessclient.models.services.error.Error;
+import com.exceptionless.exceptionlessclient.models.error.Error;
 import com.exceptionless.exceptionlessclient.queue.EventDataFilter;
 import com.exceptionless.exceptionlessclient.queue.EventValidator;
 import com.exceptionless.exceptionlessclient.utils.Utils;
