+++ src/main/java/com/exceptionless/exceptionlessclient/plugins/preconfigured/DuplicateErrorCheckerPlugin.java
@@ -3,8 +3,8 @@ package com.exceptionless.exceptionlessclient.plugins.preconfigured;
 import com.exceptionless.exceptionlessclient.configuration.ConfigurationManager;
 import com.exceptionless.exceptionlessclient.models.Event;
 import com.exceptionless.exceptionlessclient.models.EventPluginContext;
-import com.exceptionless.exceptionlessclient.models.services.error.Error;
-import com.exceptionless.exceptionlessclient.models.services.error.InnerError;
+import com.exceptionless.exceptionlessclient.models.error.Error;
+import com.exceptionless.exceptionlessclient.models.error.InnerError;
 import com.exceptionless.exceptionlessclient.plugins.EventPluginIF;
 import com.exceptionless.exceptionlessclient.plugins.MergedEvent;
 import lombok.Builder;
