+++ src/main/java/com/exceptionless/exceptionlessclient/plugins/preconfigured/ErrorPlugin.java
@@ -5,8 +5,8 @@ import com.exceptionless.exceptionlessclient.models.Event;
 import com.exceptionless.exceptionlessclient.models.EventPluginContext;
 import com.exceptionless.exceptionlessclient.models.enums.EventPropertyKey;
 import com.exceptionless.exceptionlessclient.models.enums.EventType;
-import com.exceptionless.exceptionlessclient.models.services.error.Error;
-import com.exceptionless.exceptionlessclient.models.services.error.StackFrame;
+import com.exceptionless.exceptionlessclient.models.error.Error;
+import com.exceptionless.exceptionlessclient.models.error.StackFrame;
 import com.exceptionless.exceptionlessclient.plugins.EventPluginIF;
 import lombok.Builder;
