+++ src/test/java/com/exceptionless/exceptionlessclient/plugins/preconfigured/DuplicateErrorCheckerPluginTest.java
@@ -5,9 +5,8 @@ import com.exceptionless.exceptionlessclient.configuration.ConfigurationManager;
 import com.exceptionless.exceptionlessclient.models.Event;
 import com.exceptionless.exceptionlessclient.models.EventPluginContext;
 import com.exceptionless.exceptionlessclient.models.enums.EventPropertyKey;
-import com.exceptionless.exceptionlessclient.models.services.error.Error;
-import com.exceptionless.exceptionlessclient.models.services.error.StackFrame;
-import com.exceptionless.exceptionlessclient.plugins.preconfigured.DuplicateErrorCheckerPlugin;
+import com.exceptionless.exceptionlessclient.models.error.Error;
+import com.exceptionless.exceptionlessclient.models.error.StackFrame;
 import com.exceptionless.exceptionlessclient.queue.DefaultEventQueue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
